website/requirements.txt
Flask==2.1.2
keras==2.8.0
numpy==1.22.2
opencv_contrib_python==4.5.5.64
Pillow==9.1.1
Werkzeug==2.0.3
